akhir

+++ database/migrations/2025_04_28_031312_add_product_id_to_transaction_items_table.php
@@ -1,27 +1,27 @@
 <?php
 
-use Illuminate\Database\Migrations\Migration;
-use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
-        Schema::table('transaction_items', function (Blueprint $table) {
-            $table->foreignId('product_id')->constrained()->after('transaction_id');
-        });
+        if (!Schema::hasColumn('transaction_items', 'product_id')) {
+            Schema::table('transaction_items', function (Blueprint $table) {
+                $table->foreignId('product_id')->constrained()->after('transaction_id');
+            });
+        }
     }
-    
+
     public function down(): void
     {
-        Schema::table('transaction_items', function (Blueprint $table) {
-            $table->dropForeign(['product_id']);
-            $table->dropColumn('product_id');
-        });
+        if (Schema::hasColumn('transaction_items', 'product_id')) {
+            Schema::table('transaction_items', function (Blueprint $table) {
+                $table->dropForeign(['product_id']);
+                $table->dropColumn('product_id');
+            });
+        }
     }
-    
 };
